requirements.txt
flask=1.1.1
json
requestsrequests=2.22.0
